pyproject.toml
[project]
name = "chatbot-webui"
version = "0.1.0"
description = "FastAPI backend for a ROCm-enabled local chat UI."
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<0.112",
    "uvicorn[standard]>=0.30,<0.31",
    "pydantic-settings>=2.2,<2.3",
    "sqlmodel>=0.0.21,<0.0.22",
    "alembic>=1.13,<1.14",
    "psutil>=5.9,<6.0",
    "python-multipart>=0.0.9,<0.0.10",
    "httpx>=0.27,<0.28",
    "black>=24.4,<24.5",
    "ruff>=0.5,<0.6",
    "pytest>=8.3,<8.4",
    "pytest-asyncio>=0.23,<0.24",
    "respx>=0.21,<0.22",
    "invoke>=2.2,<2.3",
]

[project.optional-dependencies]
runtime = ["llama-cpp-python>=0.2.86,<0.3"]

[tool.ruff]
line-length = 100
target-version = "py311"
lint.select = ["E", "F", "I", "UP", "B", "ASYNC"]
lint.ignore = ["B008"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["backend/tests"]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.hatch.build.targets.sdist]
include = ["backend", "README.md", "LICENSE", "implementation_plan.md"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
